v0.7: switch DSP to FunDSP hacker (f64) — fix multi-hour phase drift
Bug: after ~5 min at 48 kHz the hacker32 time counter's ULP (~3e-5) grew
larger than dt (~2e-5), so LFO phases inside lfo(|t| ...) closures began
advancing in coarse jumps instead of smooth increments. Symptom: kick and
pulse envelopes 'accelerated', reverb tails aliased, peaks spiked even
with every track muted (reverbs keep processing internally).

Fix: switch audio path to fundsp::hacker (f64) so the internal time
counter stays precise for 100+ hours.
  - All closures are now |t: f64|; Shared::value() (still f32) is cast to
    f64 at the boundary.
  - math/pulse.rs and math/sigmoid.rs converted to f64 so our shape
    functions keep pace with the closures that call them.
  - TUI code that consumes math fns (trajectory, beats) adds explicit
    casts at the boundary; TrackSnapshot stays f32 for UI display.
  - Net::get_stereo still returns (f32, f32) — only internal state moved
    to f64, so cpal interop and recording code are unchanged.
  - engine 't' accumulator moved to f64 too so phase_clock no longer
    quantises after 5 min.

Verified: 60 s headless render produces a clean 22 MB WAV.

+++ src/tui/beats.rs
@@ -13,10 +13,10 @@ const STEPS: usize = 16;
 const PHRASE_BEATS: f32 = 16.0; // 4 bars × 4 beats
 
 pub fn render(f: &mut Frame, area: Rect, engine: &EngineHandle) {
-    let t = engine.phase_clock.value();
-    let bpm = engine.global.bpm.value();
+    let t = engine.phase_clock.value() as f64;
+    let bpm = engine.global.bpm.value() as f64;
     let beat = beat_phase(t, bpm);
-    let current = (beat * STEPS as f32) as usize % STEPS;
+    let current = (beat * STEPS as f64) as usize % STEPS;
 
     // Beat row: 16 squares, current one lit, every 4th brighter baseline.
     let mut beat_spans: Vec<Span> = Vec::with_capacity(STEPS * 2 + 1);
@@ -33,8 +33,8 @@ pub fn render(f: &mut Frame, area: Rect, engine: &EngineHandle) {
     }
 
     // Phrase row: 16 small blocks for phrase progress.
-    let phr = phrase_phase(t, bpm, PHRASE_BEATS);
-    let phr_idx = (phr * STEPS as f32) as usize % STEPS;
+    let phr = phrase_phase(t, bpm, PHRASE_BEATS as f64);
+    let phr_idx = (phr * STEPS as f64) as usize % STEPS;
     let mut phrase_spans: Vec<Span> = Vec::with_capacity(STEPS * 2 + 1);
     for i in 0..STEPS {
         let (glyph, color) = if i <= phr_idx {
@@ -53,7 +53,7 @@ pub fn render(f: &mut Frame, area: Rect, engine: &EngineHandle) {
         Line::from(beat_spans),
         Line::from(""),
         Line::from(vec![Span::styled(
-            format!(" phrase  {:.0}%  ({:.0}/{:.0} beats)", phr * 100.0, phr * PHRASE_BEATS, PHRASE_BEATS),
+            format!(" phrase  {:.0}%  ({:.0}/{:.0} beats)", phr * 100.0, phr * PHRASE_BEATS as f64, PHRASE_BEATS),
             Style::default().fg(Color::Gray),
         )]),
         Line::from(phrase_spans),
